Fixed some bugs and added chain insert and create Cant seem to figure out unit Testing
