+++ src/test/java/DAO/BasicDaoTest.java
@@ -1,0 +1,15 @@
+package DAO;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class BasicDaoTest {
+
+
+    @Test
+    void drop(){
+
+    }
+
+}
