+++ src/test/java/RefelctionsWork/GetClassesTest.java
@@ -1,0 +1,7 @@
+package RefelctionsWork;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class GetClassesTest {
+
+}
